INITIAL DEV COMMIT
Screamer-Plus is an extension of constraint propagation in Screamer,
built upon a fundamental redesign of the core functions `funcallv` and `applyv`
introduced in version 4.0.1 of Screamer. This new foundation enables automatic
constraint propagation, eliminating the need for manual noticers and simplifying
function and macro definitions.

As a result, many of the macros and functions originally found in Screamer-Plus
(by Simon White)—such as `CARV`, `CDRV`, `IFV`, and others—have been entirely
rewritten or reimagined with cleaner semantics and greater efficiency.

While some function names and general ideas are inspired by the original Screamer-Plus,
all code in this package is original unless otherwise noted. Out of respect for the
community and to encourage broader collaboration, this version is released under the
permissive MIT License, in contrast to the more restrictive licensing terms of the
original Screamer-Plus.

Contributions, feedback, and extensions are warmly welcomed.